# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CyberClaw API
  description: |
    CyberClaw controlled agent platform — public API surface.

    This spec covers the **consumer-facing subset** of the full 91 routes (see
    `ROUTES.md`). Internal cluster, wizard, webhook, and dev routes are
    intentionally omitted; they are implementation details, not stable API.

    Auth: all `/api/v1/*` endpoints require `Authorization: Bearer <jwt>`.
    The token is obtained from `POST /admin/login` (HS256, `JWT_SECRET` env).
  version: 0.1.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://127.0.0.1:38090
    description: Local staging (podman, see scripts/deploy/staging-podman.sh)
  - url: https://api.cyberclaw.example.com
    description: Production (placeholder — replace with real DNS)
security:
  - bearerAuth: []
...
paths:
...
  /admin/login:
    post:
      tags: [Auth]
      summary: Issue a JWT for an existing operator
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, password]
              properties:
                user_id: { type: string, example: op_ada }
                password: { type: string }
      responses:
        '200':
          description: JWT issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  jwt:
                    type: string
                    description: "HS256-signed JWT, expires in 1h"
                  user: { $ref: '#/components/schemas/Operator' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/me:
    get:
      tags: [Auth]
      summary: Identity of the JWT holder
      responses:
        '200':
          description: Current operator
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operator' }
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Missing / invalid / expired JWT
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: unauthorized }
  schemas:
    Operator:
      type: object
      properties:
        user_id: { type: string, example: op_ada }
        role: { type: string, enum: [admin, viewer, qa] }
        display_name: { type: string }
...